@@ -12,15 +12,10 @@
 
         <!-- Scripts -->
         @vite(['resources/css/app.css', 'resources/js/app.js'])
-
-        <!-- Styles -->
-        @livewireStyles
     </head>
     <body class="font-sans antialiased">
-        <x-jet-banner />
-
         <div class="min-h-screen bg-gray-100">
-            @livewire('navigation-menu')
+            @include('layouts.navigation')
 
             <!-- Page Heading -->
             @if (isset($header))
@@ -36,9 +31,5 @@
                 {{ $slot }}
             </main>
         </div>
-
-        @stack('modals')
-
-        @livewireScripts
     </body>
 </html>
